requirements.txt
Django==2.1.3
django-celery
redis
django_celery_beat
django-mysql
mysqlclient
pyserial

